+++ myllm/handler/openai.py
@@ -11,7 +11,7 @@ from time import sleep
 from loguru import logger
 from openai import OpenAI
 
-from .client import AIClient
+from ._client import AIClient
 
 
 class OpenaiHandler(AIClient):
